backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
gunicorn==21.2.0
requests==2.31.0
whitenoise==6.6.0
python-dotenv==1.0.0
